+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.5.5
 
+### Added
+* timeout to FitHarn.preferences
+
 ## Version 0.5.4 - Released 2020-02-19 
 
 ### Added